Initial shell interface
This implements the initial admin client/shell interface (amberjack),
and renames server protocol modules.

+++ com.io7m.eigion.server.vanilla/src/main/java/com/io7m/eigion/server/vanilla/internal/EIRequestLimits.java
@@ -96,6 +96,6 @@ public final class EIRequestLimits implements EIServiceType
   public String toString()
   {
     return "[EIRequestLimits 0x%s]"
-      .formatted(Long.toUnsignedString((long) this.hashCode()));
+      .formatted(Long.toUnsignedString(this.hashCode()));
   }
 }
