+++ com.io7m.eigion.server.vanilla/src/main/java/com/io7m/eigion/server/vanilla/internal/EIServerClock.java
@@ -71,6 +71,6 @@ public final class EIServerClock implements EIServiceType
   public String toString()
   {
     return "[EIServerClock 0x%s]"
-      .formatted(Long.toUnsignedString((long) this.hashCode()));
+      .formatted(Long.toUnsignedString(this.hashCode()));
   }
 }
